Update README.md
This first project is fun and exciting.  I am going to build an application that allows nurses to parse through their union contract to find the answers to simple or complex questions that they might have regarding their rights in the workplace.  This should save them time and good will with the union reps and their managers by limiting the amount of time it takes each nurse to get answers to their questions.